fix: create a platform flow with a jdbc data store
Platform flows could not be added when configuring a JDBC data store
because of some non nullable fields being not set by the converter.

Namely this fields were:
  - id
  - createdAt
  - updatedAt

Beside this issue, the event service was trying to publish the event
with the context environment. As platform flow are not scoped to a
specific environment, the event publication was throwing an event
as well.

see https://github.com/gravitee-io/issues/issues/7399